k8s: Correctly set fsGroupPolicy for CSI driver
fsGroupPolicy defaults to ReadWriteOnceWithFSType which will limit the
use of CSI provisioned volumes because it means that fsGroup for the
PodSecurityContext can only be specified when an FSType is provided.

fsGroupPolicy of 'File' appears to function correctly in testing and
it enables us to use fsGroup in the PodSecurityContext even if the
filesystem is not specified.